Reorganize source into more modules.
